requirements.txt
Django==3.1.5
django-sslserver==0.22
djangorestframework==3.12.2
matplotlib==3.3.3
opencv-python==4.5.1.48
Pillow==8.1.0
tf-slim==1.1.0
